Move logic to new `Admin` class
